pyproject.toml
[project]
name = "d-roles"
version = "0.1.0"
description = "Gestion des rôles dans le pôle DATA"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "databases[postgresql]>=0.9.0",
    "databases>=0.7.0",
    "fastapi[standard]>=0.115.12",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
    "sqlmodel>=0.0.24",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "sentry-sdk[fastapi]>=2.29.1",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "authlib>=1.6.0",
    "requests>=2.32.4",
    "itsdangerous>=2.2.0",
    "fastapi-mail>=1.5.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
